+++ composer-c1x/src/main/java/io/piano/android/composer/c1x/C1xInterceptor.kt
@@ -1,13 +1,13 @@
 package io.piano.android.composer.c1x
 
-import com.cxense.cxensesdk.CxenseSdk
-import com.cxense.cxensesdk.model.CustomParameter
-import com.cxense.cxensesdk.model.ExternalUserId
-import com.cxense.cxensesdk.model.PageViewEvent
 import io.piano.android.composer.ExperienceInterceptor
 import io.piano.android.composer.model.ExperienceRequest
 import io.piano.android.composer.model.ExperienceResponse
 import io.piano.android.composer.model.events.ExperienceExecute
+import io.piano.android.cxense.CxenseSdk
+import io.piano.android.cxense.model.CustomParameter
+import io.piano.android.cxense.model.ExternalUserId
+import io.piano.android.cxense.model.PageViewEvent
 import timber.log.Timber
 
 class C1xInterceptor(
@@ -17,8 +17,8 @@ class C1xInterceptor(
     override fun beforeExecute(
         request: ExperienceRequest
     ) {
-        check(!request.url.isNullOrEmpty()) {
-            "URL is required for C1X"
+        check(!request.url.isNullOrEmpty() || !request.contentId.isNullOrEmpty()) {
+            "URL or Content Id is required for C1X"
         }
     }
 
@@ -40,8 +40,9 @@ class C1xInterceptor(
                     ExternalUserId(cxenseCustomerPrefix!!, it)
                 }
                 val event = PageViewEvent.Builder(
-                    siteId,
-                    request.url,
+                    siteId = siteId,
+                    location = request.url,
+                    contentId = request.contentId,
                     referrer = request.referer,
                     customParameters = mutableListOf(CustomParameter(PARAM_USERSTATE, userState)),
                 ).apply {