+++ composer/src/test/java/io/piano/android/composer/HttpHelperTest.kt
@@ -16,6 +16,7 @@ import io.piano.android.composer.model.EventExecutionContext
 import io.piano.android.composer.model.EventsContainer
 import io.piano.android.composer.model.ExperienceRequest
 import io.piano.android.composer.model.ExperienceResponse
+import io.piano.android.composer.model.UserSegmentsContainer
 import io.piano.android.composer.model.events.ShowTemplate
 import kotlin.test.Test
 import kotlin.test.assertEquals
@@ -38,7 +39,7 @@ class HttpHelperTest {
         on { isEmpty() } doReturn false
     }
     private val experienceRequest: ExperienceRequest = mock() {
-        on { customVariables } doReturn mapOf("1" to "2")
+        on { customVariables } doReturn mapOf("1" to listOf("2"))
         on { tags } doReturn listOf(DUMMY_STRING, DUMMY_STRING)
         on { customParameters } doReturn customParameters
     }
@@ -109,6 +110,7 @@ class HttpHelperTest {
                 null,
                 DUMMY_STRING2,
                 DUMMY_STRING2,
+                UserSegmentsContainer(null, null),
                 null,
                 meters
             ),