+++ id/id/src/main/java/io/piano/android/id/PianoIdActivity.kt
@@ -13,6 +13,7 @@ import android.webkit.CookieManager
 import android.webkit.WebChromeClient
 import android.webkit.WebView
 import android.webkit.WebViewClient
+import androidx.activity.OnBackPressedCallback
 import androidx.annotation.RestrictTo
 import androidx.annotation.VisibleForTesting
 import androidx.appcompat.app.AppCompatActivity
@@ -33,6 +34,7 @@ class PianoIdActivity : AppCompatActivity(), PianoIdJsInterface {
     private val jsInterface = PianoIdJavascriptDelegate(this, client.javascriptInterface)
     private var widget: String? = null
     private var disableSignUp: Boolean = false
+    private var stage: String? = null
 
     private val oauthResult = registerForActivityResult(OAuthResultContract()) {
         when (it) {
@@ -45,12 +47,19 @@ class PianoIdActivity : AppCompatActivity(), PianoIdJsInterface {
         }
     }
 
+    private val webviewBackPressedCallback = object : OnBackPressedCallback(true) {
+        override fun handleOnBackPressed() {
+            binding.webview.goBack()
+        }
+    }
+
     @SuppressLint("SetJavaScriptEnabled", "AddJavascriptInterface")
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         binding = ActivityPianoIdBinding.inflate(layoutInflater)
         setContentView(binding.root)
         intent.process()
+        onBackPressedDispatcher.addCallback(webviewBackPressedCallback)
         with(binding) {
             webview.apply {
                 savedInstanceState?.let { restoreState(it) }
@@ -80,6 +89,7 @@ class PianoIdActivity : AppCompatActivity(), PianoIdJsInterface {
                 webViewClient = object : WebViewClient() {
                     override fun onPageStarted(view: WebView, url: String?, favicon: Bitmap?) {
                         super.onPageStarted(view, url, favicon)
+                        webviewBackPressedCallback.isEnabled = view.canGoBack()
                         progressBar.show()
                     }
 
@@ -101,7 +111,7 @@ class PianoIdActivity : AppCompatActivity(), PianoIdJsInterface {
                     }
                 }
             }
-            client.getSignInUrl(disableSignUp, widget) { r ->
+            client.getSignInUrl(disableSignUp, widget, stage) { r ->
                 r.onSuccess {
                     CookieManager.getInstance().setCookie(it, "${client.aid}__ut=")
                     progressBar.isIndeterminate = false
@@ -121,14 +131,6 @@ class PianoIdActivity : AppCompatActivity(), PianoIdJsInterface {
     override fun onNewIntent(intent: Intent) {
         super.onNewIntent(intent)
         intent.process()
-    }
-
-    override fun onBackPressed() {
-        with(binding.webview) {
-            if (canGoBack())
-                goBack()
-            else super.onBackPressed()
-        }
     }
 
     override fun onSaveInstanceState(outState: Bundle) {
@@ -193,18 +195,21 @@ class PianoIdActivity : AppCompatActivity(), PianoIdJsInterface {
     internal fun Intent.process() {
         widget = getStringExtra(KEY_WIDGET)
         disableSignUp = getBooleanExtra(KEY_DISABLE_SIGN_UP, false)
+        stage = getStringExtra(KEY_STAGE)
     }
 
     companion object {
         internal const val KEY_WIDGET = "io.piano.android.id.PianoIdActivity.WIDGET"
         internal const val KEY_DISABLE_SIGN_UP = "io.piano.android.id.PianoIdActivity.DISABLE_SIGN_UP"
+        internal const val KEY_STAGE = "io.piano.android.id.PianoIdActivity.STAGE"
         internal const val JS_INTERFACE_NAME = "PianoIDMobileSDK"
 
         @JvmStatic
-        internal fun buildIntent(context: Context, disableSignUp: Boolean, widget: String?): Intent =
+        internal fun buildIntent(context: Context, disableSignUp: Boolean, widget: String?, stage: String?): Intent =
             Intent(context, PianoIdActivity::class.java)
                 .putExtra(KEY_DISABLE_SIGN_UP, disableSignUp)
                 .putExtra(KEY_WIDGET, widget)
+                .putExtra(KEY_STAGE, stage)
                 .addFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP)
     }
 }