+++ id/id/src/main/java/io/piano/android/id/PianoIdAuthResultContract.kt
@@ -11,7 +11,7 @@ class PianoIdAuthResultContract : ActivityResultContract<PianoIdClient.SignInCon
 
     override fun createIntent(context: Context, input: PianoIdClient.SignInContext): Intent =
         with(input) {
-            PianoIdActivity.buildIntent(context, disableSignUp, widget)
+            PianoIdActivity.buildIntent(context, disableSignUp, widget, stage)
         }
 
     override fun parseResult(resultCode: Int, intent: Intent?): PianoIdAuthResult? =
@@ -23,6 +23,7 @@ class PianoIdAuthResultContract : ActivityResultContract<PianoIdClient.SignInCon
                     with(intent) {
                         val code = getIntExtra(PianoId.KEY_ERROR, 0)
                         if (code == 0) {
+                            @Suppress("DEPRECATION") // deprecated since Android 13
                             PianoIdAuthResult.success(
                                 getParcelableExtra(PianoId.KEY_TOKEN),
                                 getBooleanExtra(PianoId.KEY_IS_NEW_USER, false)