+++ plugins/src/main/kotlin/io/piano/android/ktlint/KtlintConfigPlugin.kt
@@ -11,6 +11,7 @@ class KtlintConfigPlugin : Plugin<Project> {
         apply<KtlintPlugin>()
         if (this != rootProject) {
             extensions.configure(KtlintExtension::class.java) {
+                version.set("0.45.2")
                 android.set(true)
             }
         }